refactor: 시도 횟수 변수명을 number에서 count로 변경
