+++ src/view/InputView.js
@@ -1,8 +1,7 @@
 import { Console } from '@woowacourse/mission-utils';
-import throwError from '../common/utils/errorHandler.js';
-import isValidCarName from '../common/utils/carNameValidator.js';
-import isValidTryNumber from "../common/utils/tryNumberValidator.js";
-import { LOG_MESSAGE, ERROR_MESSAGE } from '../common/utils/constants/message.js';
+import { throwError } from '../common/utils.js'
+import { isValidCarName, isValidTryCount } from '../common/validator.js';
+import { LOG_MESSAGE, ERROR_MESSAGE } from '../common/constants.js';
 
 export const getUserInputCarName = async () => {
   const input = await Console.readLineAsync(LOG_MESSAGE.INPUT_CAR_NAME); 
@@ -12,9 +11,9 @@ export const getUserInputCarName = async () => {
   return input;
 }
 
-export const getUserInputTryNumber = async () => {
+export const getUserInputTryCount = async () => {
   const input = await Console.readLineAsync(LOG_MESSAGE.INPUT_TRY_NUMBER);
-  if (!isValidTryNumber(input)) {
+  if (!isValidTryCount(input)) {
     throwError(ERROR_MESSAGE.INCORRECT_TRY_NUMBER);
   }
   return input;
